refactor: 로깅 클래스 패키지 구조 변경 (#45)
